ops: moving files around for dts to be happy and build properly
