+++ package.json
@@ -1,13 +1,32 @@
 {
   "name": "@archetype-org/clack",
   "version": "0.2.0",
-  "type": "module",
   "description": "a nodejs library for sending IPC commands to an Urbit ship over Conn.c",
-  "main": "index.js",
+  "type": "module",
+  "main": "dist/index.js",
+  "module": "dist/clack.esm.mjs",
+  "types": "dist/index.d.ts",
+  "exports": {
+    ".": {
+      "require": {
+        "types": "./dist/index.d.ts",
+        "default": "./dist/index.js"
+      },
+      "import": {
+        "types": "./dist/index.d.mts",
+        "default": "./dist/clack.esm.mjs"
+      }
+    }
+  },
+  "files": [
+    "dist/**"
+  ],
   "scripts": {
+    "build": "dts build  && cp dist/index.d.ts dist/index.d.mts && cp dist/clack.esm.js dist/clack.esm.mjs",
     "test": "dts test",
     "test:watch": "dts test --watch",
-    "test:coverage": "dts test --coverage"
+    "test:coverage": "dts test --coverage",
+    "prepare": "npm run build"
   },
   "repository": {
     "type": "git",
