+++ src/khan.ts
